Schulwegsicherheit: Rename file for consistency
