+++ src/pages/XhainWayToSchool/sections/Section01Network/Map01Network.tsx
@@ -15,7 +15,7 @@ import {
   CENTER,
   MAP_STYLE,
   ZOOM,
-} from '../mapboxOptions.constants';
+} from '../mapboxOptions.const';
 import { LegendItemsNetworkSchoolways } from './LegendItemsNetworkSchoolways';
 
 export const Map01Network = () => {
