+++ src/pages/XhainWayToSchool/sections/Section02Surroundings/Map02Surroundings.tsx
@@ -15,7 +15,7 @@ import {
   CENTER,
   MAP_STYLE,
   ZOOM,
-} from '../mapboxOptions.constants';
+} from '../mapboxOptions.const';
 
 export const Map02Surroundings = () => {
   const allVisibleLayers = [
