+++ src/pages/XhainWayToSchool/sections/Section03Crossing/Map03Crossing.tsx
@@ -15,7 +15,7 @@ import {
   CENTER,
   MAP_STYLE,
   ZOOM,
-} from '../mapboxOptions.constants';
+} from '../mapboxOptions.const';
 import { LegendItemsNetworkSchoolways } from '../Section01Network/LegendItemsNetworkSchoolways';
 
 export const Map03Crossing = () => {
