+++ src/pages/ZESPlus-Research/components/MapNetworks/MapNetworks.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Map } from '~/components2/Article';
-import { BOUNDS_ZES, ZES_CENTER, ZES_ZOOM } from '../../constants';
+import { BOUNDS_ZES, ZES_CENTER, ZES_ZOOM } from '../../mapboxOptions.const';
 import {
   IconWrapper,
   Legend,
